requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.0
idna==3.4
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
requests-toolbelt==1.0.0
six==1.16.0
smartsheet-python-sdk==3.0.2
urllib3==2.0.6
